// src/app/models/recl/viewvacantes.ts
export class views {
    Id05: boolean;
    Id06: boolean;
    Id07: boolean;
    Id08: boolean;
    Id09: boolean;
    Id10: boolean;
    Id11: boolean;
    Id12: boolean;
    Id13: boolean;
    Id14: boolean;
    Id15: boolean;
    Id16: boolean;
    Id17: boolean;
    Id18: boolean;
    Id19: boolean;
    Id20: boolean;
    Id21: boolean;
    Id22: boolean;
    Id23: boolean;
    Id24: boolean;
    Id25: boolean;
    Id26: boolean;
    Id27: boolean;
    Id28: boolean;
    Id29: boolean;
    Id30: boolean;
}